+++ CHANGELOG.md
@@ -5,6 +5,10 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Added
+
+- database: `GetManagedDatabasePlans` method for retrieving available managed database plans
+
 ## [8.38.0]
 
 ### Added